Fixed errors/externs/etc for conversion to PlatformIO linking Project now compiles under PlatformIO
