+++ platformio.ini
@@ -20,3 +20,4 @@ lib_deps =
 	bakercp/CRC32@^2.0.0
 	adafruit/Adafruit HTU31D Library@^1.1.0
 	knolleary/PubSubClient@^2.8
+	jwrw/ESP_EEPROM@^2.1.1
