+++ src/parse.cpp
@@ -20,6 +20,22 @@
 #include "tasks.h"
 #include "sensor_htu31.h"
 #include "sensor_bme280.h"
+#include "main.h"
+
+void parse_system_init_device(char *, char *[], unsigned int);
+void parse_config_stub(char *, char *[], unsigned int);
+void parse_device_config_name(char *, char *[], unsigned int);
+void parse_device_config_location(char *, char *[], unsigned int);
+void parse_device_reboot(char *, char *[], unsigned int);
+void parse_device_config_wifi(char *topic, char *argv[], unsigned int argc);
+void parse_device_config_mqtt(char *topic, char *argv[], unsigned int argc);
+void parse_device_config_clear(char *topic, char *argv[], unsigned int argc);
+void parse_device_mqtt_subscribe(char *topic, char *argv[], unsigned int argc);
+void parse_debug_config_dump(char *topic, char *argv[], unsigned int argc);
+void parse_debug_rconfig_dump(char *topic, char *argv[], unsigned int argc);
+void parse_debug_get_sensor(char *topic, char *argv[], unsigned int argc);
+void parse_message(char *topic, char *omsg, unsigned int msgLength);
+
 
 struct msgCallbackList msgs[] = {
   { "SYSTEM.INIT.DEVICE", parse_system_init_device },
@@ -41,19 +57,6 @@ struct msgCallbackList msgs[] = {
   { "SET.STATE", parse_config_stub },
   { NULL, NULL },
 };
-
-void parse_system_init_device(char *, char *[], unsigned int);
-void parse_config_stub(char *, char *[], unsigned int);
-void parse_device_config_name(char *, char *[], unsigned int);
-void parse_device_config_location(char *, char *[], unsigned int);
-void parse_device_reboot(char *, char *[], unsigned int);
-void parse_device_config_wifi(char *topic, char *argv[], unsigned int argc);
-void parse_device_config_mqtt(char *topic, char *argv[], unsigned int argc);
-void parse_device_config_clear(char *topic, char *argv[], unsigned int argc);
-void parse_device_mqtt_subscribe(char *topic, char *argv[], unsigned int argc);
-void parse_debug_config_dump(char *topic, char *argv[], unsigned int argc);
-void parse_debug_rconfig_dump(char *topic, char *argv[], unsigned int argc);
-void parse_debug_get_sensor(char *topic, char *argv[], unsigned int argc);
 
 void parse_system_init_device(char *topic, char *argv[], unsigned int argc) {
   serial_printf("\nEntered parsing function for %s\n", argv[0]);
@@ -104,7 +107,7 @@ void parse_device_config_wifi(char *topic, char *argv[], unsigned int argc) {
     return;
   }
   serial_printf("parse_device_config_wifi, argc: %d\n", argc);
-  for (int i = 0; i < argc; i++) {
+  for (unsigned int i = 0; i < argc; i++) {
     serial_printf("%s ", argv[i]);
   }
   serial_printf("\n");
